test/e2e_tests/requirements.txt
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-logging
google-auth
pandas
db-dtypes
python-dotenv